Extract `useLegacyField` hook
