+++ frontend/src/metabase/common/components/ColumnValuesWidget/ColumnValuesWidget.tsx
@@ -1,9 +1,6 @@
-import { useMemo } from "react";
-import { useSelector } from "metabase/lib/redux";
-import { getMetadata } from "metabase/selectors/metadata";
 import FieldValuesWidget from "metabase/components/FieldValuesWidget";
-import * as Lib from "metabase-lib";
-import LegacyDimension from "metabase-lib/Dimension";
+import { useLegacyField } from "metabase/common/hooks/use-legacy-field";
+import type * as Lib from "metabase-lib";
 
 export interface ColumnValuesWidgetProps<T> {
   value: T[];
@@ -26,24 +23,10 @@ export function ColumnValuesWidget<T extends string | number>({
   maxWidth,
   onChange,
 }: ColumnValuesWidgetProps<T>) {
-  const metadata = useSelector(getMetadata);
-
-  const fields = useMemo(() => {
-    const fieldId = Lib._fieldId(column);
-    if (typeof fieldId === "number") {
-      const tableId = Lib._cardOrTableId(column);
-      const field = metadata.field(fieldId, tableId);
-      return field ? [field] : [];
-    }
-    const fieldRef = Lib.legacyFieldRef(column);
-    const dimension = LegacyDimension.parseMBQL(fieldRef, metadata);
-    const field = dimension?.field?.();
-    return field ? [field] : [];
-  }, [column, metadata]);
-
+  const field = useLegacyField(column);
   return (
     <FieldValuesWidget
-      fields={fields}
+      fields={field ? [field] : []}
       className="input"
       value={value}
       containerWidth="100%"
