+++ frontend/src/metabase/common/hooks/use-legacy-field/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-legacy-field";
